# Improve execution log parsing speed with CompileStatic