**Is this a bugfix, or an enhancement? Please describe.**
Improves the performance of log format line parsing.

**Describe the solution you've implemented**
`parseLine` and the other methods have been annotated with `CompileStatic` and modified to compile.

This about doubles the performance on my system with a `5mb` test file:
* ~30s -> 15s in dev mode
* ~10s -> 5s in production mode